NextWave working
Working out what the next wave will be based on the previous wave appears to be working.

Working on adding in preinstantiating all possible borls to be recycled rather than instantiating each time.

Will be changing the way the borl prefabs are made, attempted to create a nested borl with the smaller sizes inside, but it creates problems with setting the borls to inactive when they're destroyed, so they will all be independent instead.